+++ server/pkg/agent/codex_test.go
@@ -1011,13 +1011,13 @@ func TestCodexExecuteSurfacesStderrWhenChildExitsEarly(t *testing.T) {
 	}
 }
 
-func TestWithCodexStderrAppendsHint(t *testing.T) {
+func TestWithAgentStderrAppendsHint(t *testing.T) {
 	t.Parallel()
 
-	if got := withCodexStderr("codex initialize failed: process exited", ""); got != "codex initialize failed: process exited" {
+	if got := withAgentStderr("codex initialize failed: process exited", "codex", ""); got != "codex initialize failed: process exited" {
 		t.Errorf("empty tail should not modify msg, got %q", got)
 	}
-	msg := withCodexStderr("codex initialize failed: process exited", "unexpected argument '-m' found")
+	msg := withAgentStderr("codex initialize failed: process exited", "codex", "unexpected argument '-m' found")
 	want := "codex initialize failed: process exited; codex stderr: unexpected argument '-m' found"
 	if msg != want {
 		t.Errorf("got %q, want %q", msg, want)
